fix shipping settings to defalt

+++ app/services/shipping_service.rb
@@ -9,17 +9,18 @@ class ShippingService
     ceo = User.ceo
     if ceo
       admin_nation = ceo.user_nations.find_by_code(nation.code)
-      User.where(role: "user").each do |user|
-        admin_nation = ceo.user_nations.find_by_code(nation.code)
-        user_nation = user.user_nations.find_by_code(nation.code)
-        admin_nation.user_shipping_types.each do |shipping_type|
-          user_shipping_type = user_nation.user_shipping_types.find_or_create_by!(shipping_type_id: shipping_type.shipping_type.id)
-          if user_shipping_type
-            user_shipping_type.update(active: shipping_type.active)
-            admin_shipping_setting = shipping_type.shipping_settings
-            user_shipping_type.shipping_settings.delete_all
-            admin_shipping_setting.each do |shipping_setting|
-              user_shipping_type.shipping_settings.create!(min_price: shipping_setting.min_price, max_price: shipping_setting.max_price, percent: shipping_setting.percent, packet_name: shipping_setting.packet_name)
+      if admin_nation.present?
+        User.where(role: "user").each do |user|
+          user_nation = user.user_nations.find_or_create_by!(code: nation.code, name: nation.name)
+          admin_nation.user_shipping_types.each do |shipping_type|
+            user_shipping_type = user_nation.user_shipping_types.find_or_create_by!(shipping_type_id: shipping_type.shipping_type.id)
+            if user_shipping_type
+              user_shipping_type.update(active: shipping_type.active)
+              admin_shipping_setting = shipping_type.shipping_settings
+              user_shipping_type.shipping_settings.delete_all
+              admin_shipping_setting.each do |shipping_setting|
+                user_shipping_type.shipping_settings.create!(min_price: shipping_setting.min_price, max_price: shipping_setting.max_price, percent: shipping_setting.percent, packet_name: shipping_setting.packet_name)
+              end
             end
           end
         end
